Make Google Analytics tracking optional.
Some OpenStack contributors have expressed concern about having Google
Analytics tracking codes embedded in all of the HTML documentation
generated by the use of this theme. In addition to these concerns, which
mostly focus on privacy, having the Google Analytics tracking code used
widely runs the risk of creating misleading analytics data.

This change adopts an approach used by the Alabaster theme, which
enables Google Analytics tracking to be optional. Tracking is now
controlled by the theme settings in conf.py using a new theme option,
'analytics_tracking_code'. If present, this causes the HTML to include
the analytics tracking, with the provided code being used. If not
present, the Google Analytics snippet is excluded from the
documentation.

To avoid breakage, this commit uses the current hard-coded Google
Analytics tracking code as the default value of the option, which
means that most users will not notice any change and may continue to be
accidentally opted in to Google Analytics tracking.

Change-Id: I156151c3f458931f13716b4a32087afe57483794