Increase indent level of doc top contents
In the top contents of each document, it is not easy to distinguish
title levels because the current indent is 3px. It is a bit small.
This commit increases the indent level for second or more level of
titles.

Also aligns the Updated: <date> line with the page title.

margin-bottom is also unnecessary. It leads to unbalanced margin
when title level is changed.

Change-Id: I530afff51c814b96992a80a5daea8ebd81e51a53