Remove getJSON function and doc-characters file
The getJSON function amends a tag found only on the docs
landing page. When it is called on other pages it causes
a file not found error.

This patch removes the doc-characters file and the
getJSON function as they are not needed in
openstackdocstheme and they are already included in
openstack-manuals.

Closes-Bug: #1664869

Change-Id: If12287edf0a762ed0223eb2f9b33a78339a8f501