automate some of the extension setup
Remove the code that was trying to force openstackdocs to be the
theme. It was using the wrong name anyway, so it didn't work, and we
don't want to use that approach any more.

Add configuration options to make it easier to pass the git repository
and bug tracker information.

Add an event processor to update the html page context for each
page. Include the gitsha by looking at the repository. Include the
giturl by looking at the configuration option for the repo name and
building the URL from it.

Include the bug project and tag values based on the configuration
options.

Update the instructions for enabling the theme to include adding the
module to the list of extensions to turn on all of the above.

Add the openstackdocs theme automatically when the extension is
initialized.

Change-Id: Icf3a40ed104cfd828f532f6f2b112ed02f996ff5
Signed-off-by: Doug Hellmann <doug@doughellmann.com>