Serve minified CSS
My understanding of the Debian packaging requirements is that it is
acceptable to use minified builds of JS / CSS as long as the relevant
non-minified source is also supplied.

Serving minified CSS / JS can drastically reduce the size of the static
assets, reducing load times and total bandwidth usage. In some cases,
the effect of this simple process is quite significant; jquery.min, for
example, is about 32% of the size of its full size counterpart, saving
181KB on a single file.

In the future we should minify ALL static assets via toolchain, but in
the short term this is a simple and effective change.

Change-Id: I19d449c5c763e9f2e90e9137f9bfc1d21d928a15