move the localtoc for the page into the sidebar
Having a page table of contents at the top of the content section of
pages with a lot of section headings means that users have to scroll
down significantly before they see any real content. Move the page toc
into the sidebar to avoid the need for scrolling.

The old theme style information for the page toc is updated to reflect
the new location.

The example docs in demo/index.rst are updated to include nested
headings for testing the layout of the list.

Change-Id: I47222a70e58267c409c0d8d2f0f460d54c5fa5dd
Signed-off-by: Doug Hellmann <doug@doughellmann.com>