Bring back local search
This makes the search.html page that's generated by sphinx function as
the local search, and not the global google based swift_search. The
sphinx integrated search will mean that only documents linked to the
current tree will be returned, which also ensures that searching for
content in stable/pike doesn't jump you to mitaka answers. It also
brings back highlighting of search results which is a style rule that
was lost in the building of the openstackdocstheme.

We probably also want sidebar integration of this search field,
however this will at least mean hitting the search.html page will do
what people previously expected.

Change-Id: I747ad28658e12441ccad7081d15788f5d738dc5e