Sync language list with openstack-manuals repo
The language list on the top page docs.openstack.org and
the list in openstackdocstheme were out of sync.
We have eight localized index page now.

This commit updates the language list of openstackdocstheme
and changes the order of languages in the alphabetical order
which is used in the top page of docs.openstack.org.

Change-Id: I3f986ae4ea9be914cbd87a4e403479a74c8a1acc