ignore files created during packaging
The build directory is created by setuptools when a package is created.

The file RELEASENOTES.rst used to exist as a manually managed file,
and is now created when reno is installed and a package is created.

The file releasenotes/notes/reno.cache is created when reno is installed
and a package is created.

None of these files should be tracked by git, so add them to the
.gitignore list.

Change-Id: Ia5e24a9d21ed2a0e51dd81b2cd3e44231c885fcc
Signed-off-by: Doug Hellmann <doug@doughellmann.com>