Do not display "latest" badge if repo is not versioned
We do now display the badge for version even on repos that do not have
branches like https://docs.openstack.org/infra/manual . The reason is
that the test is against series - and we *always* set series, by default
to 'latest'.

Add a new option to disable showing of the badge called display_badge
and use it. Repos can set this as part of html options.

Disable showing the badge for api-ref test document.

Change-Id: I985b49412aa1848b915b8f0a2ff9e95867e420ca