Add table styling
Sphinx 2.0 switches the default HTML build from HTML4 to HTML5. HTML5
deprecates the 'border' attribute for the 'table' element and the HTML5
builder no longer emits this. This has resulted in a lack of borders on
tables.

Correct this by configuring border styling via CSS and then go a few
steps further and start colouring the table header and using zebra
stripes on rows to help visually distinguish them further. A sample
table is added to the demo documentation so we've something to inspect
locally.

Change-Id: I79b40bb5700807ac8ad523a6e0a83cd21965346e
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>
Closes-Bug: #1828265