Remove a heap of dead CSS rules
(Pronounced "hay-p" in Wesht Clare fashion).

These were discovered by searching for (and not finding) the relevant
class/id attributes in the HTML templates:

  ag -w --ignore bootstrap.min.css --ignore bootstrap.min.js \
    --ignore jquery-3.2.1.js --ignore jquery-3.2.1.min.js \
    --ignore font-awesome.min.css $term

Most of them exist from the original creation of this theme [1] and are
were either used in the past or are used in places other than the Sphinx
theme (e.g. the OpenStack website [2]). They can be removed from here.

[1] https://opendev.org/openstack/openstackdocstheme/commit/cf9c89fa9
[2] https://opendev.org/osf/openstackweb

Change-Id: Icbcbdc647d872956d150f09c37e931944d56ec2c
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>