js: Remove dead or unused JS
Remove all the commented out stuff, which is clearly not coming back,
along with code that was applied to elements with the ID of
'toggle-definitions', none of which exist. This in turns allows us to
remove all the webuiPopover stuff since that only applied to elements
with the 'gloss' class, which the 'toggle-definitions' code was
creating.

Change-Id: Ie0780e41194e5add90c8ea6f63dc496acbb0b85a
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>