Fix version dropdown
In subdirectories, the version dropdown is incorrect for the Latest
entry, it produces:
<div class="version-dropdown-content">
                    <p><a href="../latest/" rel="nofollow">Latest</a></p>
                    <p><a href="../../stein" rel="nofollow">stein</a></p>
                    <p><a href="../../rocky" rel="nofollow">rocky</a></p>

Notice the "../" instead of "../..". This breaks the "latest" link on
documents such as https://docs.openstack.org/manila/latest/admin/index.html
where "latest" links then to https://docs.openstack.org/manila/latest/latest/

Fix this by using the same logic for latest as for the other versions,
there's no need to special case it.

Change-Id: I6b0fe830f764ce7dde24cff9c7b84c6a894ca408