Improve docs title style
The style used for docs title suffers from several issues:

- the title has the same size (h2) than headers in the content
- 'last updated' info appears twice and is not *that* important
- Content starts a bit too close from the last-updated top block

This change raises the title to h1 level, and skips the top "last
updated" row unless there are several versions of the document to pick
from.

In order to keep proper separation between title and content, it
introduces a new style (title-row) and applies it to the title.

Change-Id: I8fadbe74d6a79035be987e63df649d1464cbaaa4