css: Namespace Pygments stylesheet
The 'native.css' stylesheet is the stylesheet used by Pygments. The
stylesheet defines a number of classes that map to the shortname of the
tokens generated by Pygments. By default, these classes are not
"namespaced" or prefixed by less generic selector. This is a problem
since Sphinx 3.x has started using at least one class name that
conflicts with one of the class names in Pygments, '.py'.

Sphinx itself has solved this for quite some time by doing this
"namespacing" and prepending the '.highlight' selector [1], which is the
default class name used by Pygments' 'HTMLFormatter' formatter. Do the
same thing ourselves by regenerating the Pygments stylesheet as
described in the Pygments documentation [2]:

  pygmentize -f html -S native -a .highlight > native.css

[1] https://github.com/sphinx-doc/sphinx/blob/v3.0.0/sphinx/highlighting.py#L163
[2] https://pygments.org/docs/cmdline/#generating-styles

Change-Id: I97731149d72b74fc4a92c0f5108c7a676f834fb8
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>
Closes-Bug: #1876343