Use consistent logging
Sphinx itself prefixes everything it logs with '[app]'. Historically we
had done this with '[openstackdocstheme]', but we seem to have lost our
way recently. In addition, Sphinx automatically logs everything at info
level, which means we were generating a lot of noise.

Resolve both issues by modifying our logs, adding the
'[openstackdocstheme]' prefix in the places it's missing from and lower
the logging level for those that don't need to be shown every time from
info to debug.

Change-Id: Ic116367ea7c6fc6d78b70e9c115898bae6ec6d96
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>