Added styling for rST rubric directive.
StarlingX is using the rST rubric directive to style headings such as
'Prerequisites', 'Procedure', and 'Results' in tasks without having them
appear in the TOC. There is currently no support for rubrics in the
StarlingX theme. This file appears to be the correct place to add it.
Testing to see if merge is picked up by StarlingX tox builds.

Change-Id: Ia339ad02522891c2b0b593dc782b67220565b87f
Signed-off-by: Stone <ronald.stone@windriver.com>