sphinx 4.0: Handle HTML 5 changes
Sphinx 4.0 includes a newer version of docutils, which makes some
changes to the HTML 5 writer. Among these is the change in HTML
generated for the 'ObjectDescription' directive, which is used to
generate definition-list style output for things like oslo.config.
Previously, the title of each entry would be styled like so:

  <dt id="DEFAULT.internal_service_availability_zone">
    <code class="sig-name descname">
      <span class="pre">internal_service_availability_zone</span>
    </code>
  </dt>

Now the improper use of the 'code' element is dropped, in favour of a
'span', like so:

  <dt class="sig sig-object oslo.config" id="DEFAULT.internal_service_availability_zone">
    <span class="sig-name descname">
      <span class="pre">internal_service_availability_zone</span>
    </span>
    ...
  </dt>

Change-Id: I9fb9fbce6299600c3085bea1b8f17ad796bec4f8
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>