Page header fix with Sphinx4
Previous commits addrressed some of the changes of Sphinx4 producting
different structure. The same is valid for the page heading where
instead of `<div class="section">` it produces `<section>`. This leads
to the css fix not to be appled and as such page heading present 2
times. Copy previous css fix with new path.

Change-Id: Ia9ab413280c2f98d041e0f5ae5c1df22c5a4a3f5