Fix OpenStack project "Search" pages
The OpenStack project documentations pages that implement the "Search"
capability (:ref: 'search') are not working after the recent update to
sphinx 3.4.x. The search will never end and the browser will report a
Uncaught ReferenceError: Stemmer is not defined error.
This patch updates the openstackdocstheme to fix the search capability
by including the language_data.js that now provides the "Stemmer" function.

Closes-Bug: #1957717
Change-Id: If54b3b253ed7585f6f8b20f1e0f945101ef283a3