Fix stylesheet setting in theme.conf files
The intention here was probably to include basic.css from Sphinx'
basic theme, but the correct way to reference it is just "basic.css".
This is what sphinx/themes/basic/theme.conf in Sphinx itself does.

Using "css/basic.css" caused the file to be not copied, and the
reference was 404.

Change-Id: Ifd6a78cb2dce42027f404829e1865fdb90acdb4f