Update analytics tracking ID
The company managing site analytics for docs.openstack.org has
requested we update the tracking script snippet it uses. This
switches the existing implementation to the new version they
provided to the OpenInfra Foundation staff.

A similar change I201161a51ce4483c7bad4e16fbfb39493c4ca712 covers
the copy of this in the openstack-manuals repo.

Change-Id: I91104f6f0f287e09379f89ccf3a2b084fdde9465