Fix series selection and sorting
Currently series names like 2023.1 and later appear at the end of the
version selector, but they are the most recent releases and thus should
be sorted to appear at the beginning.

Also make sure that a series doesn't appear twice in the list, which can
currently happen if e.g. the stable/victoria and unmaintained/victoria
branches exist at the same time during the transition.

Change-Id: Ide82c6de41081b9adcea632e41a581db5bc340c4